requirements.txt
asgiref==3.2.7
certifi==2020.4.5.1
chardet==3.0.4
cloudinary==1.20.0
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.0.7
django-allauth==0.41.0
django-cors-headers==3.2.1
django-heroku==0.3.1
django-rest-auth==0.9.5
django-simple-history==2.11.0
django-sslserver==0.22
djangorestframework==3.11.0
djangorestframework-camel-case==1.1.2
drf-rw-serializers==1.0.3
gunicorn==19.9.0
idna==2.9
oauthlib==3.1.0
psycopg2==2.8.5
python-dotenv==0.10.3
python3-openid==3.1.0
pytz==2020.1
requests==2.23.0
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.3.1
stripe==2.50.0
urllib3==1.25.9
whitenoise==5.1.0
